created Caesar Cipher + (Encryption/Decryption)
You have to implement a function that will decode this message and recover the original text.

The decoding process will be based on a dictionary attack, which means you have a list of words from which you expect to find the password used during the encryption process. Dictionary attacks (Links to an external site.) are one common form of password hacking method, and though this is a very small version of it, it should give you an idea of how they work. Pay close attention to the comments in the starter code!