Updated Sudoku.c
Recursion Practice with Sudoku Solver
For this exercise, you will write a straight-up (no fancy techniques) Sudoku solver. The program takes as input an array with the Sudoku to be solved, and provides the solution.

It requires recursion - so you will have to decide what the base case(s) may be, and you have to think about the recursive case that helps you find the solution.